requirements.txt
python-binance
python-dotenv
click
tabulate
